Fixing problem with DHCP Subnet Rules.
The configDhcpForSubnet() method was addedd to the Advanced Network Topology, but it should be
included in the Basic Network Topology insterad.
